fix: 解决正文图片应用lightbox的useState错误
